src: Include normal initialization routines
This change makes sure that normal initilization routines will be added to the
binary, so no weird side-effects will occur and no inline assembler to fix it
is needed. While this increases the code size by a couple of bytes, it still
remains under the 512 words limit, so its not a problem for now.